requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coverage==7.3.0
curl-cffi==0.5.9
duckduckgo-search==3.9.3
exceptiongroup==1.1.3
Faker==19.6.2
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
outcome==1.2.0
packaging==23.1
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
PyExecJS==1.5.1
pyjsparser==2.7.1
PySocks==1.7.1
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
regex==2023.10.3
requests==2.31.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
tiktoken==0.5.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
waitress==2.1.2
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.2
